+++ docs/changelog.md
@@ -10,7 +10,8 @@
 - add check for `get_filename` on output classes to override default filename
 - improved automatic `pydantic` support
 - add --timeout, --no-verify, --retries, --retry-wait options
-- fix all commands to obey various scraper options like --rpm, --header, etc.
+- add --fastmode option to use local cache
+- fix all CLI commands to obey various scraper options
 
 ## 0.7.1 - 2021-06-14